adding matcher prefix exact and regex matches for host
